Copy the rocket-chip and riscv-tools README.md
This is kind of nasty: I manually copied this stuff in and had to edit
it a bit.  I don't really want to deal with this now (blogc is super
fragile on the MD in the riscv-tools repo, I need to just fix it) so
it's what you're going to get.